requirements.txt
requests==2.27.*
pytest==7.1.*
allure-pytest==2.9.*